Add files via upload
The main goal of working on this data set is to find out what factors impact the suicide rate in the world the most. I will do that by analyzing the percentage of the suicides per gender and  age out of all countries in order to find out which cluster of the society is in the most danger. The result would describe the correlation of the suicide rate with other variables such as age, gender and population.